+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added support for reading various types of compressed files with `load_molecules`
+
+### Fixed
+- Nextflow pipeline updated to handle output types and CLI changes
+- Warnings in `load_molecules` now adhere to the CLI `--warn` argument
+
 ## [v0.3.3] - 2026-04-02
 [v0.3.3]: https://github.com/rayhackett/enzymm/compare/v0.3.2..v0.3.3